requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymysql==1.1.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
pydub==0.25.1
python-multipart==0.0.9
aiofiles==23.2.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
